+++ src/cucu/browser/selenium.py
@@ -21,6 +21,33 @@ from cucu.browser.frames import search_in_all_frames
 # suppress warning caused by selenium_keep_alive taking an extra http connection
 logging.getLogger("urllib3.connectionpool").setLevel(logging.ERROR)
 
+_DEEP_QUERY_ALL_JS = """
+var sel = arguments[0];
+var acc = [];
+function cucuWalkAll(n) {
+    if (!n) {
+        return;
+    }
+    if (n.nodeType === 1 && n.matches && n.matches(sel)) {
+        acc.push(n);
+    }
+    var ch = n.children;
+    if (ch) {
+        for (var i = 0; i < ch.length; i++) {
+            cucuWalkAll(ch[i]);
+        }
+    }
+    if (n.nodeType === 1 && n.shadowRoot) {
+        var sh = n.shadowRoot.children;
+        for (var j = 0; j < sh.length; j++) {
+            cucuWalkAll(sh[j]);
+        }
+    }
+}
+cucuWalkAll(document.documentElement);
+return acc;
+"""
+
 
 class DisableLogger:
     def __enter__(self):
@@ -334,17 +361,24 @@ class Selenium(Browser):
         return self.driver.title
 
     def css_find_elements(self, selector):
+        shadow_enabled = config.CONFIG.true("CUCU_SHADOW_DOM_SEARCH")
+
         def find_elements_in_frame():
-            elements = self.driver.find_elements(By.CSS_SELECTOR, selector)
+            if shadow_enabled:
+                elements = self.driver.execute_script(
+                    _DEEP_QUERY_ALL_JS, selector
+                )
+            else:
+                elements = self.driver.find_elements(By.CSS_SELECTOR, selector)
 
             def visible(element):
                 return element.is_displayed()
 
-            return list(filter(visible, elements))
+            return list(filter(visible, elements or []))
 
         elements = search_in_all_frames(self, find_elements_in_frame)
 
-        return elements
+        return elements or []
 
     def execute(self, javascript, *args):
         return self.driver.execute_script(javascript, *args)